requirements.txt
beautifulsoup4
html5lib
httpx
pyyaml